Update Summation to Long Type

+++ core/src/main/scala/kafka/log/Log.scala
@@ -1144,7 +1144,7 @@ class Log(@volatile var dir: File,
   /**
    * The size of the log in bytes
    */
-  def size: Long = logSegments.map(_.size).sum
+  def size: Long = logSegments.map(_.size.toLong).sum
 
   /**
    * The offset metadata of the next message that will be appended to the log
